Making b2access proxy refreshable automatically on each Eudat Endpoint call, transparent to the user.
See #39
